new file SyntaxManipulation containing utility functions for common syntax manipulations
