+++ baby-l4.cabal
@@ -75,6 +75,7 @@ library
                   KeyValueMap
                   Model
                   Syntax
+                  SyntaxManipulation
 
                   -- lsp
                   L4LSP
