+++ src/TimedMC.hs
@@ -4,10 +4,8 @@
 module TimedMC where
 
 import Syntax
-import Data.List
-import System.Console.Haskeline (putHistory)
-import Typing (funArgsToApp)
-import RuleTransfo (liftVar, conjExpr, disjExpr, abstractQ, abstractF, liftVarBy, liftType, conjsExpr)
+import SyntaxManipulation (conjExpr, disjExpr, abstractQ, abstractF, liftVarBy, conjsExpr, applyVars, mkVarE, mkEq, mkFloatConst, mkFunTp, index, indexListFromTo)
+
 import PrintProg (renameAndPrintExpr)
 
 data SMTFunDef t = SMTFunDef
@@ -36,21 +34,6 @@ TCTL formulas are composed of:
 expandEFInitial :: TA t -> Expr t -> SMTFunDef t
 expandEFInitial = undefined
 
-mkVarE :: Var t -> Expr t
-mkVarE v = VarE {annotOfExpr =  annotOfQVarName (nameOfVar v), varOfExprVarE = v}
-
-applyVars :: Var (Tp()) -> [Var (Tp())] -> Expr (Tp())
-applyVars f args = funArgsToApp (mkVarE f) (map mkVarE args)
-
-index :: Int -> Var t -> Var t
-index newi (LocalVar n _) = LocalVar n newi
-index _ v@(GlobalVar _) = v
-
-indexList :: [Int] -> [Var t] -> [Var t]
-indexList = zipWith index
-
-indexListFromTo :: Int -> Int -> [Var t] -> [Var t]
-indexListFromTo l u = indexList (reverse [l .. u])
 
 -- fnv: function variable (global)
 -- stv: state variable (local)
@@ -88,49 +71,11 @@ constrDelayTransition stv clvs =
       dgte0 = BinOpE booleanT (BCompar BCgte) (mkVarE d) (mkFloatConst 0.0)
       clockshift = conjsExpr (zipWith (\c' c -> BinOpE booleanT (BCompar BCeq) (mkVarE c') (BinOpE floatT (BArith BAadd) (mkVarE c) (mkVarE d))) (indexListFromTo 1 (n - 1) clvs) (indexListFromTo (n + 1) (2 * n - 1) clvs))
   in abstractF (sclvs ++ sclvs)
-               (conjExpr (mkCompar BCeq (index (2 * n - 1) stv) (index (n - 1) stv))
+               (conjExpr (mkEq (index (2 * n - 1) stv) (index (n - 1) stv))
                          (abstractQ Ex [d] (conjExpr dgte0 clockshift)))
 
-mkFunTp :: [Tp ()] -> Tp () -> Tp ()
-mkFunTp ts t = foldr (FunT ()) t ts
-
-mkCompar :: BComparOp -> Var (Tp()) -> Var (Tp()) -> Expr (Tp())
-mkCompar cop v1 v2 = BinOpE booleanT (BCompar cop) (mkVarE v1) (mkVarE v2)
-
-mkFloatConst :: Float -> Expr (Tp())
-mkFloatConst f = ValE floatT (FloatV f)
-
-myfnv :: String -> Var (Tp())
-myfnv f =
-  let i = integerT
-      t = ClassT () (ClsNm "Time")
-      b = booleanT
-  in GlobalVar (QVarName (mkFunTp [i, t, t, i, t, t] b)  f)
-
-mystv :: Var (Tp())
-mystv = LocalVar (QVarName integerT  "st") 0
-
-myclvs :: [Var (Tp())]
-myclvs = [LocalVar (QVarName (ClassT () (ClsNm "Time"))  "c1") 0, LocalVar (QVarName (ClassT () (ClsNm "Time"))  "c2") 0]
 
 
-constrEFStepAbstrTest :: String
-constrEFStepAbstrTest = renameAndPrintExpr [] (constrEFStepAbstr (myfnv "fn") mystv myclvs (myfnv "trans"))
-
-constrComposTest :: String
-constrComposTest = renameAndPrintExpr [] (constrCompos (mystv : myclvs) (myfnv "r1") (myfnv "r2"))
-
-constrDelayTransitionTest :: String 
-constrDelayTransitionTest = renameAndPrintExpr [] (constrDelayTransition mystv myclvs)
-
--- >>> constrDelayTransitionTest
--- "( \\ st: Integer -> ( \\ c1: Time -> ( \\ c2: Time -> ( \\ st_0: Integer -> ( \\ c1_0: Time -> ( \\ c2_0: Time -> ((st@5==st_0@2)&&( exists d: Float. ((d@0>=0.0)&&((c1_0@2==(c1@5+d@0))&&(c2_0@1==(c2@4+d@0))))))))))))"
-
-
-{-
-
-
--}
 
 expandEFStep :: String -> Int -> TA t -> SMTAbstr t
 expandEFStep = undefined
@@ -152,3 +97,38 @@ mcEFbounded k ta phi = do
     let expansion = expandEF k 0 ta "f" [expandEFInitial ta phi]
     checkExpansion expansion
 -}
+
+-- BEGIN for testing
+mystv :: Var (Tp())
+mystv = LocalVar (QVarName integerT  "st") 0
+
+myclvs :: [Var (Tp())]
+myclvs = [LocalVar (QVarName (ClassT () (ClsNm "Time"))  "c1") 0, LocalVar (QVarName (ClassT () (ClsNm "Time"))  "c2") 0]
+
+myfnv :: String -> Var (Tp())
+myfnv f =
+  let i = integerT
+      t = ClassT () (ClsNm "Time")
+      b = booleanT
+  in GlobalVar (QVarName (mkFunTp [i, t, t, i, t, t] b)  f)
+
+constrEFStepAbstrTest :: String
+constrEFStepAbstrTest = renameAndPrintExpr [] (constrEFStepAbstr (myfnv "fn") mystv myclvs (myfnv "trans"))
+
+constrComposTest :: String
+constrComposTest = renameAndPrintExpr [] (constrCompos (mystv : myclvs) (myfnv "r1") (myfnv "r2"))
+
+constrDelayTransitionTest :: String 
+constrDelayTransitionTest = renameAndPrintExpr [] (constrDelayTransition mystv myclvs)
+
+-- >>> constrDelayTransitionTest
+-- "( \\ st: Integer -> ( \\ c1: Time -> ( \\ c2: Time -> ( \\ st_0: Integer -> ( \\ c1_0: Time -> ( \\ c2_0: Time -> ((st@5==st_0@2)&&( exists d: Float. ((d@0>=0.0)&&((c1_0@2==(c1@5+d@0))&&(c2_0@1==(c2@4+d@0))))))))))))"
+
+
+{-
+
+
+-}
+
+
+-- END for testing
